refactor(gateway): harvest shared media-type helpers from dingtalk and wecom into SDK
